Homework/HW4/requirements.txt
transformers==4.51.3
datasets==2.21.0
accelerate==1.6.0
peft==0.15.2
rouge==1.0.1
torchmetrics==1.7.1